Adding hold_timer to peers for timing keepalives and ending stale sessions
In order to keep track of the freshness of peer sessions, adding support for hold_timers in the config. Peer hold_timer is sent in the OPEN, and session is instantiated with the lesser of the local and remote hold_timer value. In order to support this, these features were added:
- HoldTimer struct that keeps track of the last update, when keepalives should be sent (via Future.poll()), and the remaining hold_time for a peer
- Elapsed time formatting for logging the Session
- An additional poll in the Session.poll() that checks for the need to send a keepalive (or a hold time expiration)