Select! Cleanup (#7)
Using the new tokio::select! macro from 0.2.11 to cleanup some usages of select! (removing fuse and pin)

There was previously a 250ms timeout workaround in the session (to not block the manager) that can now go away as the hold_timer .tick() will bump the Session stream, allowing manager to move forward.